requirements.txt
tensorflow==2.15.0
numpy
pandas
matplotlib
scikit-learn
pillow
opencv-python
librosa
soundfile
tqdm
flask
